# docs/phase1/api/openapi.yaml
openapi: 3.0.3
info:
  title: NovelShelf API
  description: 個人利用の広告なしWeb小説リーダー「NovelShelf」のバックエンドAPI。
  version: 0.1.0-phase1
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /shelf:
    get:
      tags: [shelf]
      summary: 本棚一覧取得
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ShelfStatus" }
        - name: favorite
          in: query
          schema: { type: boolean }
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [author, series, site, tag]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BookshelfEntry" }
    post:
      tags: [shelf]
      summary: 作品を本棚に登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novelId, status]
              properties:
                novelId: { type: string, format: uuid }
                status: { $ref: "#/components/schemas/ShelfStatus" }
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookshelfEntry" }
  /shelf/{entryId}:
    patch:
      tags: [shelf]
      summary: 本棚エントリのステータス/お気に入り/タグを更新
      parameters:
        - $ref: "#/components/parameters/EntryId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: "#/components/schemas/ShelfStatus" }
                isFavorite: { type: boolean }
                tagIds:
                  type: array
                  items: { type: string, format: uuid }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookshelfEntry" }
    delete:
      tags: [shelf]
      summary: 本棚から削除
      parameters:
        - $ref: "#/components/parameters/EntryId"
      responses:
        "204": { description: 成功 }

  # ---------- Reading position / history ----------
...
  /stats/breakdown:
    get:
      tags: [stats]
      summary: 月別/サイト別/作者別の読書時間内訳
      parameters:
        - name: by
          in: query
          required: true
          schema:
            type: string
            enum: [month, site, author]
        - name: range
          in: query
          schema:
            type: string
            enum: [today, week, month, year, all]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StatsBreakdownItem" }
  /stats/calendar:
    get:
      tags: [stats]
      summary: 読書カレンダー（日別読書時間）
      parameters:
        - name: yearMonth
          in: query
          required: true
          schema: { type: string, example: "2026-07" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CalendarDay" }

  # ---------- Updates ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EntryId:
      name: entryId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    ShelfStatus:
      type: string
      enum: [READING, COMPLETED, READ_LATER]
    Novel:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        author: { type: string }
        site: { type: string, enum: [NAROU, KAKUYOMU, HAMELN, PIXIV] }
        genre: { type: string }
        coverUrl: { type: string, format: uri }
        sourceUrl: { type: string, format: uri }
        status: { type: string, enum: [ONGOING, COMPLETED] }
        latestKnownChapterNo: { type: integer }
        hasUpdate: { type: boolean }
...
    BookshelfEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        novel: { $ref: "#/components/schemas/Novel" }
        status: { $ref: "#/components/schemas/ShelfStatus" }
        isFavorite: { type: boolean }
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
        addedAt: { type: string, format: date-time }
...
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
...
    StatsBreakdownItem:
      type: object
      properties:
        label: { type: string }
        readingSeconds: { type: integer }
    CalendarDay:
      type: object
      properties:
        date: { type: string, format: date }
        readingSeconds: { type: integer }
...